moved registerServiceWorker to src/utils/
