+++ src/App.jsx
@@ -151,4 +151,4 @@ const mapDispatchToProps = dispatch => ({
 
 })
 
-export default App;
+export default connect(mapStateToProps, mapDispatchToProps)(App)
